+++ examples/data-sources/gpcloud_project/data-source.tf
@@ -1,0 +1,3 @@
+resource "gpcloud_project" "example" {
+  id = "b194dd6f-21ea-44e0-98b7-b7ac434430d3"
+}