+++ CHANGELOG.md
@@ -1,9 +1,23 @@
-## 0.1.1 (Unreleased)
+## 0.1.2 (Beta)
+
+Added:
+- New Authentication method: Service Account (Client Credentials)
+- Added Data Source `gpcloud_project` for read-only access to projects
+
+Changed:
+- Updated resource and datasource documentation
+- Updated README.md
+
+Fixed:
+- Updated the terraform-plugin-sdk from 2.24.1 to 2.25.0
+
+
+## 0.1.1 (Beta)
 
 SPELLING:
 - Rename Flavor to Flavour to represent the same spelling as the API
 
-## 0.1.0 (Unreleased)
+## 0.1.0 (Beta)
 
 FEATURES:
 - Initial release of the provider